Lazy registration for all methods (#313)
For packages, we now register methods twice: once at build-time (so code in your package works) and once on load (so that user code works). This includes updating `new_external_generic()` so that it works for S3 and S4 generics, and renaming `external_methods_register()` to `methods_register()`.

Part of #306.